test: harden compiled shape guard integration
