+++ crates/tenferro-runtime/src/shape_infer.rs
@@ -509,10 +509,7 @@ fn require_input<'a>(
     })
 }
 
-pub(crate) fn resolve_dim_expr_from_shapes(
-    expr: &DimExpr,
-    input_shapes: &[&[DimExpr]],
-) -> Result<DimExpr> {
+fn resolve_dim_expr_from_shapes(expr: &DimExpr, input_shapes: &[&[DimExpr]]) -> Result<DimExpr> {
     match expr {
         DimExpr::Const(value) => Ok(DimExpr::Const(*value)),
         DimExpr::InputDim { input_idx, axis } => {
